+++ libraries/session/src/main/java/androidx/media3/session/MediaController.java
@@ -388,6 +388,9 @@ public class MediaController implements Player {
      * changes the available commands} for a controller that affect whether buttons of the custom
      * layout are enabled or disabled.
      *
+     * <p>Note that the {@linkplain CommandButton#isEnabled enabled} flag is set to {@code false} if
+     * the available commands do not allow to use a button.
+     *
      * @param controller The controller.
      * @param layout The ordered list of {@linkplain CommandButton command buttons}.
      */
@@ -975,6 +978,9 @@ public class MediaController implements Player {
    *
    * <p>After being connected, a change of the custom layout is reported with {@link
    * Listener#onCustomLayoutChanged(MediaController, List)}.
+   *
+   * <p>Note that the {@linkplain CommandButton#isEnabled enabled} flag is set to {@code false} if
+   * the available commands do not allow to use a button.
    *
    * @return The custom layout.
    */