+++ libraries/session/src/main/java/androidx/media3/session/MediaControllerImplBase.java
@@ -120,7 +120,8 @@ import org.checkerframework.checker.nullness.qual.NonNull;
   private boolean released;
   private PlayerInfo playerInfo;
   @Nullable private PendingIntent sessionActivity;
-  private ImmutableList<CommandButton> customLayout;
+  private ImmutableList<CommandButton> customLayoutOriginal;
+  private ImmutableList<CommandButton> customLayoutWithUnavailableButtonsDisabled;
   private SessionCommands sessionCommands;
   private Commands playerCommandsFromSession;
   private Commands playerCommandsFromPlayer;
@@ -146,7 +147,8 @@ import org.checkerframework.checker.nullness.qual.NonNull;
     playerInfo = PlayerInfo.DEFAULT;
     surfaceSize = Size.UNKNOWN;
     sessionCommands = SessionCommands.EMPTY;
-    customLayout = ImmutableList.of();
+    customLayoutOriginal = ImmutableList.of();
+    customLayoutWithUnavailableButtonsDisabled = ImmutableList.of();
     playerCommandsFromSession = Commands.EMPTY;
     playerCommandsFromPlayer = Commands.EMPTY;
     intersectedPlayerCommands =
@@ -726,7 +728,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 
   @Override
   public ImmutableList<CommandButton> getCustomLayout() {
-    return customLayout;
+    return customLayoutWithUnavailableButtonsDisabled;
   }
 
   @Override
@@ -2611,8 +2613,9 @@ import org.checkerframework.checker.nullness.qual.NonNull;
     intersectedPlayerCommands =
         createIntersectedCommandsEnsuringCommandReleaseAvailable(
             playerCommandsFromSession, playerCommandsFromPlayer);
-    customLayout =
-        CommandButton.getEnabledCommandButtons(
+    customLayoutOriginal = result.customLayout;
+    customLayoutWithUnavailableButtonsDisabled =
+        CommandButton.copyWithUnavailableButtonsDisabled(
             result.customLayout, sessionCommands, intersectedPlayerCommands);
     playerInfo = result.playerInfo;
     try {
@@ -2765,6 +2768,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
     if (!playerCommandsChanged && !sessionCommandsChanged) {
       return;
     }
+    this.sessionCommands = sessionCommands;
     boolean intersectedPlayerCommandsChanged = false;
     if (playerCommandsChanged) {
       playerCommandsFromSession = playerCommands;
@@ -2776,13 +2780,12 @@ import org.checkerframework.checker.nullness.qual.NonNull;
           !Util.areEqual(intersectedPlayerCommands, prevIntersectedPlayerCommands);
     }
     boolean customLayoutChanged = false;
-    if (sessionCommandsChanged) {
-      this.sessionCommands = sessionCommands;
-      ImmutableList<CommandButton> oldCustomLayout = customLayout;
-      customLayout =
-          CommandButton.getEnabledCommandButtons(
-              customLayout, sessionCommands, intersectedPlayerCommands);
-      customLayoutChanged = !customLayout.equals(oldCustomLayout);
+    if (sessionCommandsChanged || intersectedPlayerCommandsChanged) {
+      ImmutableList<CommandButton> oldCustomLayout = customLayoutWithUnavailableButtonsDisabled;
+      customLayoutWithUnavailableButtonsDisabled =
+          CommandButton.copyWithUnavailableButtonsDisabled(
+              customLayoutOriginal, sessionCommands, intersectedPlayerCommands);
+      customLayoutChanged = !customLayoutWithUnavailableButtonsDisabled.equals(oldCustomLayout);
     }
     if (intersectedPlayerCommandsChanged) {
       listeners.sendEvent(
@@ -2798,7 +2801,9 @@ import org.checkerframework.checker.nullness.qual.NonNull;
     if (customLayoutChanged) {
       getInstance()
           .notifyControllerListener(
-              listener -> listener.onCustomLayoutChanged(getInstance(), customLayout));
+              listener ->
+                  listener.onCustomLayoutChanged(
+                      getInstance(), customLayoutWithUnavailableButtonsDisabled));
     }
   }
 
@@ -2816,10 +2821,23 @@ import org.checkerframework.checker.nullness.qual.NonNull;
             playerCommandsFromSession, playerCommandsFromPlayer);
     boolean intersectedPlayerCommandsChanged =
         !Util.areEqual(intersectedPlayerCommands, prevIntersectedPlayerCommands);
+    boolean customLayoutChanged = false;
     if (intersectedPlayerCommandsChanged) {
+      ImmutableList<CommandButton> oldCustomLayout = customLayoutWithUnavailableButtonsDisabled;
+      customLayoutWithUnavailableButtonsDisabled =
+          CommandButton.copyWithUnavailableButtonsDisabled(
+              customLayoutOriginal, sessionCommands, intersectedPlayerCommands);
+      customLayoutChanged = !customLayoutWithUnavailableButtonsDisabled.equals(oldCustomLayout);
       listeners.sendEvent(
           /* eventFlag= */ Player.EVENT_AVAILABLE_COMMANDS_CHANGED,
           listener -> listener.onAvailableCommandsChanged(intersectedPlayerCommands));
+    }
+    if (customLayoutChanged) {
+      getInstance()
+          .notifyControllerListener(
+              listener ->
+                  listener.onCustomLayoutChanged(
+                      getInstance(), customLayoutWithUnavailableButtonsDisabled));
     }
   }
 
@@ -2829,19 +2847,24 @@ import org.checkerframework.checker.nullness.qual.NonNull;
     if (!isConnected()) {
       return;
     }
-    ImmutableList<CommandButton> oldCustomLayout = customLayout;
-    customLayout =
-        CommandButton.getEnabledCommandButtons(layout, sessionCommands, intersectedPlayerCommands);
-    boolean hasCustomLayoutChanged = !Objects.equals(customLayout, oldCustomLayout);
+    ImmutableList<CommandButton> oldCustomLayout = customLayoutWithUnavailableButtonsDisabled;
+    customLayoutOriginal = ImmutableList.copyOf(layout);
+    customLayoutWithUnavailableButtonsDisabled =
+        CommandButton.copyWithUnavailableButtonsDisabled(
+            layout, sessionCommands, intersectedPlayerCommands);
+    boolean hasCustomLayoutChanged =
+        !Objects.equals(customLayoutWithUnavailableButtonsDisabled, oldCustomLayout);
     getInstance()
         .notifyControllerListener(
             listener -> {
               ListenableFuture<SessionResult> future =
                   checkNotNull(
-                      listener.onSetCustomLayout(getInstance(), customLayout),
+                      listener.onSetCustomLayout(
+                          getInstance(), customLayoutWithUnavailableButtonsDisabled),
                       "MediaController.Listener#onSetCustomLayout() must not return null");
               if (hasCustomLayoutChanged) {
-                listener.onCustomLayoutChanged(getInstance(), customLayout);
+                listener.onCustomLayoutChanged(
+                    getInstance(), customLayoutWithUnavailableButtonsDisabled);
               }
               sendControllerResultWhenReady(seq, future);
             });