+++ libraries/test_session_current/src/androidTest/java/androidx/media3/session/MediaControllerCompatPlaybackStateCompatActionsWithMediaSessionTest.java
@@ -1446,10 +1446,9 @@ public class MediaControllerCompatPlaybackStateCompatActionsWithMediaSessionTest
     Player player = createDefaultPlayer();
     Bundle extras1 = new Bundle();
     extras1.putString("key1", "value1");
-    Bundle extras2 = new Bundle();
-    extras1.putString("key2", "value2");
     SessionCommand command1 = new SessionCommand("command1", extras1);
-    SessionCommand command2 = new SessionCommand("command2", extras2);
+    SessionCommand command2 = new SessionCommand("command2", Bundle.EMPTY);
+    SessionCommand command3 = new SessionCommand("command3", Bundle.EMPTY);
     ImmutableList<CommandButton> customLayout =
         ImmutableList.of(
             new CommandButton.Builder()
@@ -1461,6 +1460,12 @@ public class MediaControllerCompatPlaybackStateCompatActionsWithMediaSessionTest
                 .setDisplayName("button2")
                 .setIconResId(R.drawable.media3_notification_pause)
                 .setSessionCommand(command2)
+                .build(),
+            new CommandButton.Builder()
+                .setDisplayName("button3")
+                .setEnabled(false)
+                .setIconResId(R.drawable.media3_notification_pause)
+                .setSessionCommand(command3)
                 .build());
     MediaSession.Callback callback =
         new MediaSession.Callback() {
@@ -1469,7 +1474,11 @@ public class MediaControllerCompatPlaybackStateCompatActionsWithMediaSessionTest
               MediaSession session, MediaSession.ControllerInfo controller) {
             return new AcceptedResultBuilder(session)
                 .setAvailableSessionCommands(
-                    ConnectionResult.DEFAULT_SESSION_COMMANDS.buildUpon().add(command1).build())
+                    ConnectionResult.DEFAULT_SESSION_COMMANDS
+                        .buildUpon()
+                        .add(command1)
+                        .add(command3)
+                        .build())
                 .build();
           }
         };